Increased Performance and Zero-Indexing
By using a small "static" lookup table on the BitSet class and iterating
over the words instead of the bits, the performance for cardinality was
increased by upwards of 100% in my tests.

The stated goal of the class (line 1) is to be compatable with the Java
implementation, which is zero-indexed. Using zero-indexing also stays in
line with JavaScript convention, which is always zero-indexed. Many of
the tests were changed to reflect this.

This commit also brings the bitset package up-to-date with its
dependencies. The mocha package included a breaking change related to
parsing CoffeeScript, which required an update to the Cakefile.